% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.r
\name{auc}
\alias{auc}
\title{A function to compute area under a curve.}
\usage{
auc(xrange, mod, dense.N, taus, auc.version = "response")
}
\arguments{
\item{xrange}{y}

\item{mod}{x}

\item{dense.N}{the number of areas to cut into in the calculation of area under the curve; default = 201}

\item{taus}{determine the output the percentile of env variable; default = c(0,95,100) for response.}

\item{auc.version}{calculation version, "response" or "tolerance"; default = "response".}
}
\value{
returns area under the specified curve.
}
\description{
Used in other functions in this library (taxon.response(), taxon.response.sort(), and tolerance()).
Different versions for response and tolerance.
}
\examples{
xrange <- 1
lrm1 <- 2

# response version

dense.N <- 201
taus=c(0,95,100)
#' lrm1.95f <- auc.response(xrange, lrm1, dense.N, taus, "response")
# tolerance version

taus = c(50,95)
lrm195f <- auc(xrange, lrm1, dense.N, taus, "tolerance")
}
\keyword{area}
\keyword{curve,}
\keyword{under}

