% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve.shape.r
\name{curve.shape}
\alias{curve.shape}
\title{A function to determine the curve shape of a binomial probability}
\usage{
curve.shape(mnr, ubnd, lbnd)
}
\arguments{
\item{mnr}{mean predicted probability}

\item{ubnd}{upper boundary of mean predicted probability}

\item{lbnd}{lower boundary of mean predicted probability}
}
\value{
returns one of "unimodal", "Increasing", "Decreasing", "Concave up", or NA.
}
\description{
Adapted from Lester Yuan's code
}
\examples{
up.bound <- exp(up.bound.link)/(1 + exp(up.bound.link))
low.bound <- exp(low.bound.link)/(1 + exp(low.bound.link))
mean.resp <- exp(mean.resp.link)/(1 + exp(mean.resp.link))
tolcl[i] <- curve.shape(mean.resp, up.bound,low.bound)
}
\keyword{cdf,}
\keyword{gam,}
\keyword{hc05,}
\keyword{logistic}
\keyword{quantiles,}
\keyword{regression,}
\keyword{response}
\keyword{taxon}
\keyword{xc95,}

