% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{df1}
\alias{df1}
\title{df1}
\format{a data frame with 228 variables:}
\usage{
df1
}
\description{
Species crosstab (ss) with environmental variable lgSO4HCO3 and lat/long fields.
}
\keyword{datasets}
