% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.check.add.r
\name{dir.check.add}
\alias{dir.check.add}
\title{A function to check for a directory and add if it doesn't exist.}
\usage{
dir.check.add(Dir.main = getwd(), Dir.sub)
}
\arguments{
\item{Dir.main}{Main/root directory; default is getwd()}

\item{Dir.sub}{sub directory to check for and add if necessary.}
}
\value{
Nothing to return
}
\description{
Used in several functions before saving output.  Ensures that the code doesn't fail by making sure the directory exists.
}
\examples{
Dir.main <- getwd()
Dir.sub <- "output"
dir.check.add(Dir.main, Dir.sub)
}
\keyword{add,}
\keyword{check,}
\keyword{create}
\keyword{directory,}
