% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{dta.do}
\alias{dta.do}
\title{dta.do}
\format{A data frame with 3,259 obs. and 68 variables:}
\usage{
dta.do
}
\description{
For fish.weighted.cdf() example, environmental data (DO).
}
\keyword{datasets}

