% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{envdata.all}
\alias{envdata.all}
\title{envdata.all}
\format{A data frame with 518 obs. and 16 variables:}
\usage{
envdata.all
}
\description{
For taxon.response() example, environmental data.
}
\keyword{datasets}

