% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.gam.r
\name{fish.gam}
\alias{fish.gam}
\title{GAM}
\usage{
fish.gam(datafile1, datafile2, ss1, ss2, taxon, SampleID = "FishInvertFID",
  wtHUC = TRUE, SampleID2 = "Station_Date", fit = TRUE, np = 61,
  nt = 25)
}
\arguments{
\item{datafile1}{Environmental data file 1.}

\item{datafile2}{Environmental data file 2.}

\item{ss1}{Species data file 1.}

\item{ss2}{Species data file 2.}

\item{taxon}{Species data frame taxon name.}

\item{SampleID}{Site/sample id column; default = "FishInvertFID"}

\item{wtHUC}{Should data be weighted by HUC; default = TRUE}

\item{SampleID2}{Sample id field; default = "Station_Date"}

\item{fit}{= TRUE}

\item{np}{Number of bins; default = 61.}

\item{nt}{Minimum number of occurence; default = 25.}
}
\value{
**Need something here**
}
\description{
GAM
}
\examples{
#
}
\keyword{cdf,}
\keyword{gam,}
\keyword{hc05,}
\keyword{logistic}
\keyword{quantiles,}
\keyword{regression,}
\keyword{response}
\keyword{taxon}
\keyword{xc95,}

