% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.wt.cdf.r
\name{fish.wt.cdf}
\alias{fish.wt.cdf}
\title{Weighted cdf function to calculate XC95 values}
\usage{
fish.wt.cdf(datafile = datafile, ss = ss, plot = F, dogam = F,
  SampleID = "Station_Date", tag = "", sortvect = NULL, np = 61,
  nt = 25, addtrend = F)
}
\arguments{
\item{datafile}{environmental data, default = "datafile" from global environment.}

\item{ss}{Species crosstabed data; default = "ss" from global environment.}

\item{plot}{A boolean to choose if plot cdf and gam plots; default = F.}

\item{dogam}{A booleen to choose if a gam fit is calculated; default = F.}

\item{SampleID}{Site/sample id column; default = "Station_Date"}

\item{tag}{Default = "".}

\item{sortvect}{to provide a vector of species list so plots will be sorted according to the list; default = NULL.}

\item{np}{Number of bins; default = 61.}

\item{nt}{Minimum number of occurence; default = 25.}

\item{addtrend}{A booleen if a trend should be added ( = ">" etc) in the output (T or F); default = F.}
}
\value{
A dataframe of XC95 values.
}
\description{
Weighted cdf function to calculate XC95 values
}
\keyword{cdf,}
\keyword{gam,}
\keyword{hc05,}
\keyword{logistic}
\keyword{quantiles,}
\keyword{regression,}
\keyword{response}
\keyword{taxon}
\keyword{xc95,}

