% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{species}
\alias{species}
\title{species}
\format{A data frame with 14,747 obs. and 26 variables; SampID and x taxa::}
\usage{
species
}
\description{
For taxon.response() example, a species table of relative abundance of each taxon at each site/sample ID.
}
\keyword{datasets}
