% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ss}
\alias{ss}
\title{ss}
\format{A data frame with 225 variables:
\describe{
\item{\code{Sample.ID}}{sample/site ID}
}}
\usage{
ss
}
\description{
Species crosstab table of relative abundance of each taxon at each site/sample ID (row).
}
\keyword{datasets}
