% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ss.sites}
\alias{ss.sites}
\title{ss.sites}
\format{A list with 211 with 3789 levels for Station_Date::}
\usage{
ss.sites
}
\description{
For fish.weighted.cdf() example, a species list of relative abundance of each taxon at each site/sample ID.
}
\keyword{datasets}

