% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.response.r
\name{taxon.response}
\alias{taxon.response}
\title{Plot Taxa Response Curve and Calculate Species Extirpation Concentrations}
\usage{
taxon.response(spdata, envdata, sp.siteid = "Sample.ID", species = "GENUS",
  sp.abndid = "RA", env.siteid = "Sample.ID", xvar = "COND",
  cutoff = 30, region = "all", lim = "CDF", coord = NULL, mtype = 3,
  dense.N = 201, plot.pdf = F, add.map = F, statename = NULL,
  add.lab = F,
  main = "Capture Probability of Macroinvertebrate Taxon Along Conductivity Gradient",
  mar = c(5, 4, 3, 2), xlabs = expression(paste("Conductivity ( ", mu,
  "S/cm)")), log.x = TRUE, rounder = 0, taus = c(0, 95, 100), nbin = 61)
}
\arguments{
\item{spdata}{Species data.}

\item{envdata}{Environmental data.}

\item{sp.siteid}{Site/sample id column; default = "Sample.ID"}

\item{species}{; default = "GENUS"}

\item{sp.abndid}{; default = "RA"}

\item{env.siteid}{; default = "Sample.ID"}

\item{xvar}{; default = "COND"}

\item{cutoff}{; default = 30}

\item{region}{; default = "all"}

\item{lim}{; default ="CDF"}

\item{coord}{; default = NULL}

\item{mtype}{; default = 3}

\item{dense.N}{; default = 201}

\item{plot.pdf}{; default = F}

\item{add.map}{; default =F}

\item{statename}{; default = NULL}

\item{add.lab}{; default = F}

\item{main}{Plot title (main); default = "Capture Probability of Macroinvertebrate Taxon Along Conductivity Gradient"}

\item{mar}{; default = c(5,4,3, 2)}

\item{xlabs}{; default = expression(paste("Conductivity ( ", mu, "S/cm)"))}

\item{log.x}{; default = TRUE}

\item{rounder}{xvar rounder, default = 0.}

\item{taus}{; default = c(0,95,100)}

\item{nbin}{Number of bins; default = 61.}
}
\value{
**Need something here**
}
\description{
The output of this function to return 1.Weighted Average, 2. cdf_Abundance based, 3. cdf_ presence/absence based;
4. ecdf weighted, 5. cdf weight new; 6. Linear logistic regression, 7. quadratic logistic 8. GAM  5~7 using full data range;
9~11. repeat 6~8 but uses observed range for each single taxon; 12 Count. 13. Raw quantiles
}
\examples{
#
}
\keyword{cdf,}
\keyword{gam,}
\keyword{hc05,}
\keyword{logistic}
\keyword{quantiles,}
\keyword{regression,}
\keyword{response}
\keyword{taxon}
\keyword{xc95,}

