% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tol.env.cond}
\alias{tol.env.cond}
\title{tol.env.cond}
\format{A data frame with 482 obs. and 128 variables:}
\usage{
tol.env.cond
}
\description{
For tolerance() example, environmental data.
}
\keyword{datasets}
