% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tol.ss}
\alias{tol.ss}
\title{tol.ss}
\format{A data frame with 1095 obs. and 340 variables; SampID and 229 taxa::}
\usage{
tol.ss
}
\description{
For tolerance() example, a species crosstab table of relative abundance of each taxon at each site/sample ID (row).
}
\keyword{datasets}

