% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightcdf2.r
\name{weightcdf}
\alias{weightcdf}
\title{Simple weighted cdf}
\usage{
weightcdf(df1 = my.samp, ss = ss, SampID = "Sample.ID", xvar = "cond",
  nt = 25, log = TRUE, np = 61)
}
\arguments{
\item{df1}{; default = my.samp}

\item{ss}{Species crosstabed data; default = "ss" from global environment.}

\item{SampID}{Site/sample id column; default = "Sample.ID"}

\item{xvar}{variable on which to base calculations; default  = "cond"}

\item{nt}{Minimum number of occurence; default = 25.}

\item{log}{A boolean for if the values should be log (base 10) transformed; default = TRUE}

\item{np}{Number of bins; default = 61.}
}
\value{
data frame: taxaname, XC95.all, N
}
\description{
Uses wtd.quantile() from Hmisc.
}
\examples{
### Step 1A. use samples to calcualte HC05
### basic models for season, and ecoregion
allyear <- weightcdf(df1 = bio.sample[!is.na(bio.sample$lgSO4HCO3),], ss = ss, SampID = "Sample.ID", xvar = "lgSO4HCO3", nt = 25)
}
\keyword{cdf,}
\keyword{gam,}
\keyword{hc05,}
\keyword{logistic}
\keyword{quantiles,}
\keyword{regression,}
\keyword{response}
\keyword{taxon}
\keyword{xc95,}
