% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve.shape.r
\name{curve.shape}
\alias{curve.shape}
\title{A function to determine the curve shape of a binomial probability}
\usage{
curve.shape(mnr, ubnd, lbnd)
}
\arguments{
\item{mnr}{mean predicted probability}

\item{ubnd}{and lbnd upper and lower boundaries of mean predicted probability}
}
\value{
return one of "unimodal", "Increasing", "Decreasing", "Concave up", or NA.
}
\description{
adapted from Lester Yuan's code
}
\examples{
curve.shape(mean.resp, up.bound,low.bound)
}
\keyword{logistic}
\keyword{quantiles,}
\keyword{regression,}

