% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve.shape.r
\name{curve.shape}
\alias{curve.shape}
\title{A function to determine the curve shape of a binomial probability}
\usage{
curve.shape(mnr, ubnd, lbnd)
}
\arguments{
\item{mnr}{mean predicted probability}

\item{ubnd}{upper boundary of mean predicted probability}

\item{lbnd}{lower boundary of mean predicted probability}
}
\value{
returns one of "Unimodal", "Increasing", "Decreasing", "Concave up", or NA.
}
\description{
Adapted from Lester Yuan's code.  Used in taxon.response.sort().
}
\examples{
mean.resp <- 1:1000
up.bound <- mean.resp+1
low.bound <- mean.resp-1
curve.shape(mean.resp, up.bound,low.bound)
}
\keyword{cdf,}
\keyword{gam,}
\keyword{hc05,}
\keyword{logistic}
\keyword{quantiles,}
\keyword{regression,}
\keyword{response}
\keyword{taxon}
\keyword{xc95,}

