% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.response.r
\name{taxon.response}
\alias{taxon.response}
\title{Plot Taxa Response Curve and Calculate Species Extirpation Concentrations}
\usage{
taxon.response(spdata, envdata, sp.siteid = "Sample.ID", species = "GENUS",
  sp.abndid = "RA", env.siteid = "Sample.ID", xvar = "COND",
  cutoff = 30, region = "all", lim = "CDF", coord = NULL, mtype = 3,
  dense.N = 201, plot.pdf = F, add.map = F, statename = NULL,
  add.lab = F,
  main = "Capture Probability of Macroinvertebrate Taxon Along Conductivity Gradient",
  mar = c(5, 4, 3, 2), xlabs = expression(paste("Conductivity ( ", mu,
  "S/cm)")), log.x = TRUE, rounder = 0, taus = c(0, 95, 100), nbin = 61)
}
\arguments{
\item{cutoff}{a required minimum sample size for calculation}

\item{region}{a subregion code to name the final output files}

\item{lim}{if lim == "GAM", add gam plot xc95 otherwise, add   "CDF"}

\item{mtype}{could be 1 to 3, indicating which regression model to use.}

\item{dense.N}{is the number of areas to cut into in the calculation of area under the curve}

\item{plot.pdf}{to decide if we want species vs. env plots options "none", "pdf", "tiff"}

\item{add.map}{to decide if a map should be added before plots.}

\item{log.x}{if xvar should be logtransformated}

\item{rounder}{xvar rounder, default is 0}

\item{taus}{determine the output the percentile of env variable}

\item{nbin}{number of bins for logits}

\item{df1, }{data frame}

\item{sort.vect}{when plot, sort the taxa list according to a vector called file called sort.vec}

\item{xvar.}{xvariable, could be column index or name}

\item{maintext}{title of the multiplots area}
}
\value{
**Need something here**
}
\description{
The output of this function to return 1.Weighted Average, 2. cdf_Abundance based, 3. cdf_ presence/absence based;
4. ecdf weighted, 5. cdf weight new; 6. Linear logistic regression, 7. quadratic logistic 8. GAM  5~7 using full data range; 
9~11. repeat 6~8 but uses observed range for each single taxon; 12 Count. 13. Raw quantiles
}
\examples{
#**Need something here**
}
\keyword{logistic}
\keyword{quantiles,}
\keyword{regression,}

