% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance.r
\name{tolerance}
\alias{tolerance}
\title{Tolerance}
\usage{
tolerance(spdata, envdata, sp.siteid = "Sample.ID", species = "GENUS",
  covar = NULL, sp.abndid = "RA", env.siteid = "Sample.ID",
  xvar = "cond", cutoff = 20, cutoff2 = 10, region = "all",
  lim = "GAM", coord = NULL, mtype = 3, dense.N = 201, cast = TRUE,
  plot.pdf = F, add.map = F, statename = NULL, add.lab = F,
  add.abund = T,
  main = "Capture Probability of Macroinvertebrate Taxon Along Conductivity Gradient",
  mar = c(5, 4, 3, 4), xlabs = expression(paste("Conductivity ( ", mu,
  "S/cm)")), log.x = TRUE, plus = F, rounder = 0, taus = c(50, 95),
  nbin = 61)
}
\arguments{
\item{spdata}{Species data.}

\item{envdata}{Environmental data.}

\item{sp.siteid}{Species data frame column name for site/sample id; default = "Sample.ID"}

\item{species}{Species data frame column name for taxa name; default = ""GENUS"}

\item{covar}{; default = NULL}

\item{sp.abndid}{Species data frame column name for species abundance; default = "RA"}

\item{env.siteid}{Environmental data frame column name for site/sample id.  Default is "Sample.ID"}

\item{xvar}{Environmental data frame column name for analyte used in analysis; default = "cond"}

\item{cutoff}{; default = 20}

\item{cutoff2}{; default = 10}

\item{region}{; default = "all"}

\item{lim}{"GAM" or "CDF" ; default = "GAM"  if lim == "GAM", add gam plot xc95 otherwise, add   "CDF"}

\item{coord}{c("LONG_DD", "LAT_DD").  Default is NULL.}

\item{mtype}{1 linear, 2 quadratic 3 gam model for ploting purpose; default = 3.}

\item{dense.N}{the number of areas to cut into in the calculation of area under the curve; default = 201.}

\item{cast}{; default = TRUE,}

\item{plot.pdf}{to decide if we want specise vs. env plots; default = F}

\item{add.map}{boolean for if a map should be plotted.  Requires statename coord; default = F}

\item{statename}{State name used for map; default = NULL}

\item{add.lab}{boolean for adding labels to map; default = F}

\item{add.abund}{boolean for adding abundance to map; default = T.}

\item{main}{= "Capture Probability of Macroinvertebrate Taxon Along Conductivity Gradient",}

\item{mar}{Plot margins; default = c(5,4,3, 4)}

\item{xlabs}{; default = expression(paste("Conductivity ( ", mu, "S/cm)"))}

\item{log.x}{To decide if we want to log transform and plot data; default = TRUE.  log.x to use log1o transformed data, plus to use log10(+1) transformed data, sqrt}

\item{plus}{; default = F}

\item{rounder}{How many digits; default = 0.}

\item{taus}{determine the output the percentile of env variable; default = c(50,95)}

\item{nbin}{number of bins; default = 61}
}
\value{
**Need something here**
}
\description{
A function to calculate species response curves using various models.
Get optimum value from taxon-environment relationship
1.WA, 2. cdf_Abundance 3. cdf_p/a. 4. ecdf weighted
5. Linear logistic 6. quadratic logistic 7. GAM  5~7 using full data range;
 11. Count. 12. Raw quantiles
 Requires the "reshape" library.
}
\examples{
#
}
\keyword{cdf,}
\keyword{gam,}
\keyword{hc05,}
\keyword{logistic}
\keyword{quantiles,}
\keyword{regression,}
\keyword{response}
\keyword{taxon}
\keyword{xc95,}

